Normalized intensities in figure generated by runDensityExample
Shading of subplots in figure grid produced by runDensityExample are
now based on common minima and maxima.  By default, the five percent
of wStar elements that are most extreme are considered outliers.  (The
elements of *all* wStar matrices are considered in aggregate; any
given wStar matrix may be free of outliers, or may contain many.)  The
grayscale mapping shows the least non-outlier value as black, and the
greatest non-outlier value as white.  The outliers are mapped to these
clipped extrema.