Added policy options for generation of dense non-source W
Under the new policies, the non-source W (referred to in the code
mostly as auxW) has normally distributed random elements and is
normalized by row, column, or determinant, depending on the auxWPolicy
option.  Scaling of auxW is supported through the auxWWeight option.
By default, the normalization is absolute under the determinant
policy, and relative to W under the row and column policies.  These
defaults may be overridden with a suffix to the auxWPolicy option.

Added several error/sanity checks.