Fixed Opts.optConflict; added param 'noise' for both source and dest.
Separate noise params still exist for source and dest, but the generic
'noise' param can be used to set the source and dest noise to the same
value.  Distinct source and dest noise values can be specified with
'noiseAtSource' and 'noiseAtDest' (cf. 'noisinessForSource' and
'noisinessForDest' in earlier versions).

With the defaults in this commit, the invocation

  optcell={'auxWPolicy','beta','lizierNorm',false,'zScoreSigs',true,'noise',1};
  CausalSimulator.runDensityExample(optcell{:});

yields an output grid in which the W*'s closely resemble the W's.
(That is significant mainly as a baseline sanity test.  Note that the
W's themselves are not depicted in the output.  The W-like entities
depicted are the nonsource-W's and the W*'s.)