Changed 'isDestBalancing'; added option 'preWPolicy' for source input
The preWPolicy allows the source to be treated as a kind of
destination in its own right (so as to make source and destination
less dissimilar qualitatively).

New function CausalSimulator.baselineRunDensity approximately
reproduces baseline behavior of CausalSimulator.runDensityExample
of recent commits (with parameters adjusted for baseline).

With this commit, the invocation

  WClassifier.baselineExample;

yields the following scores, which are slightly different from the
baseline scores at the previous commit.  The differences may be due to
changes in evaluation order, resulting in different random values (but
this conjecture has not been verified).

Scores:
  Columns 1 through 8

    1.0000    1.0000    0.9500    1.0000    0.8000    1.0000    1.0000    0.9250

  Columns 9 through 10

    1.0000    1.0000

Mean = 0.9675