Reconciled Pipeline and new multivariate GrangerCausality
Reconciliation entailed changes both to Pipeline.m and to
GrangerCausality.m.  Pipeline encountered three kinds of problems when
using the new GrangerCausality:

1. aclags too low, causing MATLAB indexing errors;
2. aclags too high, causing excessive running time;
3. negative determinant of SIG(x,x) or SIGR(x,x), causing complex GCs.

This commit bounds aclags to address problems 1 and 2, and to address
problem 3, returns a zero GC when either determinant is non-positive.
There may be better ways of dealing with these problems.

Added diagnostic provisions to assist with debug/test.