Added preliminary implementation of accuracy plots
Static method constructTestPlotData invokes makePlotCapsule to create
sample plot-data file, which can then be rendered as a plot by
makePlotFromCapsule.  The idea is to separate generation of the plot
data, which is computation-intensive and can be done in batch mode
without graphics, from the plotting itself.

Because of problems with the causality computations, among others, the
data generated and plotted by the current prototype is partly fudged.