Results of exploratory runs of plot generation
This commit includes Pipeline.m as of Feb 22, 2015, at 21:09, just
before the initiation of a run of constructTestPlotData.  As revised
for this commit, constructTestPlotData generates data for four plots
to show the effects of varying four different opt parameters.

This commit also includes the plot-capsule files generated by the run
of constructTestPlotData initiated following the code revisions.