More results of exploratory runs of plot generation
This commit includes Pipeline.m as of Feb 23, 2015, at 20:54, just
before the initiation of a run of constructTestPlotData.  As revised
for this commit, constructTestPlotData generates data for two plots to
show the effects of varying nRun and WFullness.

In these runs, WSum is set to 0.25.

Like the previous commit, this commit also includes the plot-capsule
files generated by the run of constructTestPlotData initiated
following the code revisions.