Corrections to previous commit (also: error in commit-comment noted)
Previous commit-comment incorrectly asserted that last night's runs
used WSum=0.25.  That value for WSum appeared in the Pipeline
constructor usage comments, but the value for WSum in the ParseArgs
call remained at 0.1, as for the four runs of the previous night.

This commit changes Pipeline.m so that a new set of six runs this
evening will all use a WSum value of 0.2 (except where varied).