Changed behavior of normVar, including addition of new normVar choice
The new normVar=1 behavior preserves the proportions that the CRecur,
W, and noise terms contribute in the recurrence for a given signal.
That is, with normVar=1, normalization is achieved by multiplying the
entire right-hand side by a variance-normalizing factor, not by
individually changing any of the coefficients on the right-hand side.

The former normVar=1 behavior, slightly modified, is now obtained with
normVar=2.

Also revised seeSigVariances.m to show effect of all normVar choices.